+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ### Features
 
+- v3.0.0
+
+  - Major internal rewrite to remove ReactDOMServer dependency.
+
+  - Using react hooks internally, so support for React versions before hooks is now dropped. Use `v2.0.1` if you want to use this library in versions before React@16.8 (pre-hooks)
+
+  **Breaking:**
+
+  - `icon` option changed to `renderIcon`, where you can pass a render function instead of a node. (Useful for Lazy Rendering)
+  - Export for `create` function removed. `cogoToast()` works like create did before.
+
 - v2.0.1
 
   - Fix for top level typings declaration